Mark input snapshotting finished in execution engine
Moving more parts over to execution engine.
